requirements.txt
streamlit==1.32.0
python-dotenv==1.0.0
opencv-python==4.8.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
google-cloud-vision==3.4.5
google-generativeai==0.3.1
openpyxl==3.1.2